+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Version 0.1.1 (2019-04-24)
 * enhancement: minor fixes to examples and drop redundant examples
 * feature: implement pair/type T/FRIN/FROUT with m/l/r in Utils.type_side_to_inout
 * enhancement: add reference to Buliga write-up on beta re-writes in GraphRewriter.js
+* feature: add example `predecessor(one)`
 
 
 Version 0.1.0 (2019-04-24)